app/requirements2.txt
flask
moviepy
opencv-python
boto3
openai
nltk
matplotlib
Pillow==9.5.0